Refactor negotiation sniffer into modular submodules
